requirements.txt
openai~=1.75.0
python-dotenv~=1.1.0
streamlit~=1.44.1